+++ packages/core-js/proposals/map-update-or-insert.js
@@ -1,1 +1,2 @@
-require('../modules/esnext.map.update-or-insert');
+// TODO: remove from `core-js@4`
+require('./map-upsert');