+++ packages/core-js/proposals/map-upsert.js
@@ -1,0 +1,4 @@
+// https://github.com/thumbsupep/proposal-upsert
+// TODO: remove from `core-js@4`
+require('../modules/esnext.map.update-or-insert');
+require('../modules/esnext.map.upsert');