+++ tests/tests/index.js
@@ -267,6 +267,7 @@ import './esnext.map.reduce';
 import './esnext.map.some';
 import './esnext.map.update';
 import './esnext.map.update-or-insert';
+import './esnext.map.upsert';
 import './esnext.set.add-all';
 import './esnext.set.delete-all';
 import './esnext.set.difference';